# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/example-presets": {
      "get": {
        "operationId": "ExamplePresetController_getExamplePresets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "예제 프리셋 목록을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExamplePresetsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "example-presets"
        ]
      },
      "post": {
        "operationId": "ExamplePresetController_createExamplePreset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamplePresetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "예제 프리셋을 생성합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExamplePresetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "example-presets"
        ]
      }
    },
    "/example-presets/{id}": {
      "get": {
        "operationId": "ExamplePresetController_getExamplePresetDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프리셋 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예제 프리셋 상세 정보를 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExamplePresetDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "example-presets"
        ]
      },
      "delete": {
        "operationId": "ExamplePresetController_deleteExamplePreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프리셋 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예제 프리셋을 삭제합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteExamplePresetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "example-presets"
        ]
      },
      "patch": {
        "operationId": "ExamplePresetController_updateExamplePreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프리셋 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExamplePresetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "예제 프리셋을 수정합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateExamplePresetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "example-presets"
        ]
      }
    },
    "/example-presets/{id}/examples": {
      "get": {
        "operationId": "ExamplePresetController_getExamplePresetExamples",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프리셋 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "sourceLanguage",
            "required": true,
            "in": "query",
            "description": "예제를 조회할 소스 언어",
            "schema": {
              "type": "string",
              "enum": [
                "ko",
                "en",
                "ja",
                "zh"
              ]
            }
          },
          {
            "name": "targetLanguage",
            "required": true,
            "in": "query",
            "description": "예제를 조회할 타겟 언어",
            "schema": {
              "type": "string",
              "enum": [
                "ko",
                "en",
                "ja",
                "zh"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 언어 쌍에 대한 예제 문장을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExamplePresetExamplesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "example-presets"
        ]
      }
    },
    "/example-presets/load": {
      "post": {
        "operationId": "ExamplePresetController_loadExamplePreset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadExamplePresetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "예제 프리셋을 로드합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadExamplePresetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "example-presets"
        ]
      }
    },
...
  },
  "info": {
    "title": "FormattedDocs API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ExamplePresetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "프리셋 ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "프리셋 이름",
            "example": "기본 프리셋"
          },
          "description": {
            "type": "object",
            "description": "프리셋 설명",
            "example": "이것은 기본 프리셋입니다.",
            "nullable": true
          },
          "languages": {
            "description": "프리셋에 포함된 언어 목록",
            "example": [
              "ko",
              "en"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "languages"
        ]
      },
      "GetExamplePresetsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "presets": {
            "description": "예제 프리셋 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamplePresetDto"
            }
          },
          "currentPreset": {
            "type": "string",
            "description": "현재 선택된 프리셋 이름",
            "example": "기본 프리셋"
          }
        },
        "required": [
          "success",
          "presets",
          "currentPreset"
        ]
      },
      "ExamplePresetDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "프리셋 ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "프리셋 이름",
            "example": "기본 프리셋"
          },
          "description": {
            "type": "object",
            "description": "프리셋 설명",
            "example": "이것은 기본 프리셋입니다.",
            "nullable": true
          },
          "languages": {
            "description": "프리셋에 포함된 언어 목록",
            "example": [
              "ko",
              "en"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "languages"
        ]
      },
      "GetExamplePresetDetailResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "preset": {
            "description": "예제 프리셋 상세 정보",
            "allOf": [
              {
                "$ref": "#/components/schemas/ExamplePresetDetailDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "preset"
        ]
      },
      "ExamplePresetExampleLineDto": {
        "type": "object",
        "properties": {
          "sourceText": {
            "type": "string",
            "description": "원문 예제 문장",
            "example": "안녕하세요"
          },
          "resultText": {
            "type": "string",
            "description": "번역 예제 문장",
            "example": "Hello"
          }
        },
        "required": [
          "sourceText",
          "resultText"
        ]
      },
      "GetExamplePresetExamplesResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "sourceLanguage": {
            "type": "string",
            "description": "예제를 조회한 소스 언어",
            "enum": [
              "ko",
              "en",
              "ja",
              "zh"
            ]
          },
          "targetLanguage": {
            "type": "string",
            "description": "예제를 조회한 타겟 언어",
            "enum": [
              "ko",
              "en",
              "ja",
              "zh"
            ]
          },
          "examples": {
            "description": "요청한 언어 쌍에 해당하는 예제 문장 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamplePresetExampleLineDto"
            }
          }
        },
        "required": [
          "success",
          "sourceLanguage",
          "targetLanguage",
          "examples"
        ]
      },
      "LoadExamplePresetRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "로드할 프리셋 이름",
            "example": "기본 프리셋"
          }
        },
        "required": [
          "name"
        ]
      },
      "LoadExamplePresetResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
      "CreateExamplePresetRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "새 프리셋 이름",
            "example": "나만의 프리셋"
          },
          "description": {
            "type": "object",
            "description": "프리셋 설명",
            "example": "이것은 나만의 프리셋입니다.",
            "nullable": true
          },
          "examples": {
            "type": "object",
            "description": "언어별 번역 예제 목록",
            "example": {
              "Korean": {
                "English": {
                  "sourceLines": [
                    "안녕하세요"
                  ],
                  "resultLines": [
                    "Hello"
                  ]
                }
              }
            }
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "CreateExamplePresetResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "preset": {
            "description": "생성된 예제 프리셋 정보",
            "allOf": [
              {
                "$ref": "#/components/schemas/ExamplePresetDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "preset"
        ]
      },
      "DeleteExamplePresetResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
      "UpdateExamplePresetRequestDto": {
        "type": "object",
        "properties": {
          "examples": {
            "type": "object",
            "description": "언어별 번역 예제 목록",
            "example": {
              "Korean": {
                "English": {
                  "sourceLines": [
                    "안녕하세요"
                  ],
                  "resultLines": [
                    "Hello"
                  ]
                }
              }
            }
          },
          "description": {
            "type": "object",
            "description": "프리셋 설명",
            "example": "이것은 업데이트된 프리셋입니다.",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "프리셋 이름",
            "example": "업데이트된 프리셋"
          }
        },
        "required": [
          "examples"
        ]
      },
      "UpdateExamplePresetResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}